requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chatbot==1.5.2b0
click==8.1.8
colorama==0.4.6
comm==0.2.2
comtypes==1.4.11
contourpy==1.3.0
cvzone==1.6.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
executing==2.1.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.19.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.9.0
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.14.0
h5py==3.14.0
httpcore==1.0.7
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.5
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.34
jaxlib==0.4.34
jedi==0.19.2
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.11.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.1.0
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.4
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.1.2
openai==1.107.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.17.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.1
psycopg2==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
pyjokes==0.8.3
PyJWT==2.10.1
pymongo==4.11.3
pyparsing==3.2.0
pypiwin32==223
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.2.1
python-multipart==0.0.20
pyttsx==1.1
pyttsx3==2.99
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2025.9.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.1.0
rpds-py==0.22.3
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.14.1
Send2Trash==1.8.3
sentencepiece==0.2.1
setuptools==75.7.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.1
statistic==0.1.6
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.0
torch==2.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
twilio==9.8.0
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.3.0
utils==1.0.2
uvicorn==0.34.0
virtualenv==20.34.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.3
yarl==1.20.1
